Use row-major order for matrix operations
Previously, my multiply function assumed matrices
were in row-major order but the transformation
functions assumed column-major.

Fix it so everything uses row-major order as I
find this most natural (it’s like maths). However,
we need to be careful in the vertex shader because
it assumes column-major.

Fortunately, we can post-multiply the matrix and
everything Just Works.