Calculate the normals of geometric models
The VertexNormals system iterates through all
entities that have geometry that don’t have
Normals set on them.

It then calculates normals for the geometric model
and stores these in BufferData on another entity.
The new entity is associated with both the
geometric model and the instance of that model.