Incorporate materials into the lighting model
When an a_color is also specified, multiply it by
the material’s color - a kind of blending.
Otherwise, a_color can be set to (1.0, 1.0, 1.0)
so it doesn’t interfere with the material.

The specular highlights aren’t multiplied by
v_color as this approximates the reflection from
the light so this is probably more realistic.